+++ include/device/gemm/native_2dmm.cuh
@@ -1,3 +1,3 @@
 # include <cuda_runtime.h>
 
-__global__ void matrix_multiply_cuda_naive(float *a, float *b, float *c, int M, int N,int K);
+__global__ void matrix_multiply_cuda_naive(int *a, int *b, int *c, int M, int N,int K);
